+++ src/2DAffine.cpp
@@ -40,14 +40,15 @@ struct _2DAffine : Module {
 	float x2 = 0.f;
 	float x10 = 0.f;
 	float x20 = 0.f;
-	float y1 = 0.f;
-	float y2 = 0.f;
+	float y1[16] = {};
+	float y2[16] = {};
 	float Theta = 0.f;
 	float DeltaTheta = 0.f;
 	float xoff_pre = 0.f;
 	float yoff_pre = 0.f;
 	float xoff_post = 0.f;
 	float yoff_post = 0.f;
+	int   panelTheme;
 
 	_2DAffine() {
 		config(NUM_PARAMS, NUM_INPUTS, NUM_OUTPUTS, NUM_LIGHTS);
@@ -62,44 +63,51 @@ struct _2DAffine : Module {
 		configParam(YOFF_PRE_PARAM, -10.f, 10.f, 0.0f, "Y_Offset");
 		configParam(XOFF_POST_PARAM, -10.f, 10.f, 0.0f, "X_Offset");
 		configParam(YOFF_POST_PARAM, -10.f, 10.f, 0.0f, "Y_Offset");
+		panelTheme = 0;
 	}
 
 	void process(const ProcessArgs& args) override {
-		// Read input values
-		x1   = inputs[IN1_INPUT].getVoltage();
-		x2   = inputs[IN2_INPUT].getVoltage();
-		xoff_pre = params[XOFF_PRE_PARAM].getValue();
-		yoff_pre = params[YOFF_PRE_PARAM].getValue();
-		x1 = x1 + xoff_pre;
-		x2 = x2 + yoff_pre;
 
-		// Scale x
-		x10 = x1 + x2 * (params[SHEARX_PARAM].getValue() + inputs[CVSHEARX_INPUT].getVoltage() * params[SHEARXAMOUNT_PARAM].getValue());
-			
-		// Shear y
-		x20 = x2 + x1 * (params[SHEARY_PARAM].getValue() + inputs[CVSHEARY_INPUT].getVoltage() * params[SHEARYAMOUNT_PARAM].getValue());
-		x1 = x10; 
-		x2 = x20;
+		// Use the first input to get number of channels
+		int channels = std::max(inputs[IN1_INPUT].getChannels(), 1);
 
 		// Define rotation angle
 		Theta = params[ANGLE_PARAM].getValue() / 180 * M_PI;
 		DeltaTheta = M_PI / 5.0 * inputs[CV_INPUT].getVoltage() * params[AMOUNT_PARAM].getValue();
 		Theta += DeltaTheta;
 
-		// Rotation
-		y1 =   std::cos(Theta) * x1 + std::sin(Theta) * x2;
-		y2 = - std::sin(Theta) * x1 + std::cos(Theta) * x2;	
+		// Define translation parameters 
+		xoff_pre = params[XOFF_PRE_PARAM].getValue();
+		yoff_pre = params[YOFF_PRE_PARAM].getValue();
 		xoff_post = params[XOFF_POST_PARAM].getValue();
 		yoff_post = params[YOFF_POST_PARAM].getValue();
-	
-		//outputs[OUT1_OUTPUT].setVoltage(clamp(y1+xoff_post,-10.f,10.f));
-		//outputs[OUT2_OUTPUT].setVoltage(clamp(y2+yoff_post,-10.f,10.f));
-		// Parabolic saturation
-		y1 = (y1 + xoff_post)/10.f;
-		y2 = (y2 + yoff_post)/10.f;
-		outputs[OUT1_OUTPUT].setVoltage(10.f * clamp(y1,-2.f,2.f) * (1.f-0.25*std::abs(y1)));
-		outputs[OUT2_OUTPUT].setVoltage(10.f * clamp(y2,-2.f,2.f) * (1.f-0.25*std::abs(y2)));
+		
 
+		for (int c = 0; c < channels; c++) {		
+			// Read input values
+			x1   = inputs[IN1_INPUT].getVoltage(c);
+			x2   = inputs[IN2_INPUT].getVoltage(c);
+			x1 = x1 + xoff_pre;
+			x2 = x2 + yoff_pre;
+
+			// Scale x & y
+			x10 = x1 + x2 * (params[SHEARX_PARAM].getValue() + inputs[CVSHEARX_INPUT].getVoltage() * params[SHEARXAMOUNT_PARAM].getValue());
+			x20 = x2 + x1 * (params[SHEARY_PARAM].getValue() + inputs[CVSHEARY_INPUT].getVoltage() * params[SHEARYAMOUNT_PARAM].getValue());
+
+			// Rotation
+			y1[c] = (   std::cos(Theta) * x10 + std::sin(Theta) * x20 + xoff_post)/10.f;
+			y2[c] = ( - std::sin(Theta) * x10 + std::cos(Theta) * x20 + yoff_post)/10.f;	
+		
+			// Parabolic saturation
+			y1[c] = 10.f * clamp(y1[c],-2.f,2.f) * (1.f-0.25*std::abs(y1[c]));
+			y2[c] = 10.f * clamp(y2[c],-2.f,2.f) * (1.f-0.25*std::abs(y2[c]));
+		}
+
+		// Set output values
+		outputs[OUT1_OUTPUT].setChannels(channels);
+		outputs[OUT2_OUTPUT].setChannels(channels);
+		outputs[OUT1_OUTPUT].writeVoltages(y1);
+		outputs[OUT2_OUTPUT].writeVoltages(y2);
 
 		//Light
 		lights[XOFF_PRE_LIGHT].setBrightness(std::abs(xoff_pre)/10);
@@ -107,12 +115,73 @@ struct _2DAffine : Module {
 		lights[XOFF_POST_LIGHT].setBrightness(std::abs(xoff_post)/10);
 		lights[YOFF_POST_LIGHT].setBrightness(std::abs(yoff_post)/10);
 	}
+
+
+	json_t *dataToJson() override {
+		json_t *rootJ = json_object();
+		// panelTheme
+		json_object_set_new(rootJ, "panelTheme", json_integer(panelTheme));
+		return rootJ;
+	}
+
+	
+	void dataFromJson(json_t *rootJ) override {
+		// panelTheme
+		json_t *panelThemeJ = json_object_get(rootJ, "panelTheme");
+		if (panelThemeJ)
+			panelTheme = json_integer_value(panelThemeJ);
+		//resetNonJson();
+	}
+
 };
 
 
 struct _2DAffineWidget : ModuleWidget {
+
+	SvgPanel* darkPanel;
+
+	struct PanelThemeItem : MenuItem {
+		_2DAffine *module;
+		int theme;
+		void onAction(const event::Action &e) override {
+			module->panelTheme = theme;
+		}
+		void step() override {
+			rightText = (module->panelTheme == theme) ? "✔" : "";
+		}
+	};	
+	void appendContextMenu(Menu *menu) override {
+		MenuLabel *spacerLabel = new MenuLabel();
+		menu->addChild(spacerLabel);
+
+		_2DAffine *module = dynamic_cast<_2DAffine*>(this->module);
+		assert(module);
+
+		MenuLabel *themeLabel = new MenuLabel();
+		themeLabel->text = "Panel Theme";
+		menu->addChild(themeLabel);
+
+		PanelThemeItem *lightItem = new PanelThemeItem();
+		lightItem->text = "Light panel";
+		lightItem->module = module;
+		lightItem->theme = 0;
+		menu->addChild(lightItem);
+
+		PanelThemeItem *darkItem = new PanelThemeItem();
+		darkItem->text = "Dark panel";	
+		darkItem->module = module;
+		darkItem->theme = 1;
+		menu->addChild(darkItem);
+	}	
+
 	_2DAffineWidget(_2DAffine* module) {
 		setModule(module);
+		if (module) {
+			darkPanel = new SvgPanel();
+			darkPanel->setBackground(APP->window->loadSvg(asset::plugin(pluginInstance, "res/2DAffine_dark.svg")));
+			darkPanel->visible = false;
+			addChild(darkPanel);
+		}
 		setPanel(APP->window->loadSvg(asset::plugin(pluginInstance, "res/2DAffine.svg")));
 
 		addChild(createWidget<ScrewSilver>(Vec(0, 0)));
@@ -141,6 +210,16 @@ struct _2DAffineWidget : ModuleWidget {
 		addOutput(createOutputCentered<PJ301MPort>(mm2px(Vec(7.533, 118.046)),  module, _2DAffine::OUT1_OUTPUT));
 		addOutput(createOutputCentered<PJ301MPort>(mm2px(Vec(17.795, 118.046)), module, _2DAffine::OUT2_OUTPUT));
 	}
+
+	void step() override {
+		if (module) {
+			panel->visible = ((((_2DAffine*)module)->panelTheme) == 0);
+			darkPanel->visible  = ((((_2DAffine*)module)->panelTheme) == 1);
+		}
+		Widget::step();
+	}
+
+
 };