+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+#### 1.0.2 (2020-01-25)
+- **All Modules**: Add support for polyphony
+- **All Modules**: Add option in the context menu for light or dark panel
+- **Polygonal VCO**: Add option in the context menu to quantize the N parameter
+
 #### 1.0.1 (2020-01-14)
-- **2DAffine**: Add parabolic saturation at the output as the shearing may increase a lot the signal dynamic.
+- **2DAffine**: Add parabolic saturation at the output as the shearing may increase a lot the signal dynamic
 
 - **MarokSeq**: 
 	1. Add output triggers for each state
@@ -8,6 +13,5 @@
 	4. Add a context menu to choose the behavior if all probabilities are 0
 
  
-
 #### 1.0.0 (2020-01-08)
-- Initial Release of 2DAffine, 2DRotation, MarkovSeq, PolginalVCO.
+- Initial Release of 2DAffine, 2DRotation, MarkovSeq, PolygonalVCO